# src/WC.Service.Employees.API/WC.Service.Employees.API.swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "NSwag.AspNetCore.Launcher",
    "version": "v1"
  },
  "paths": {
    "/api/v1/colleagues": {
      "get": {
        "tags": [
          "Colleague"
        ],
        "summary": "Retrieves a list of colleagues.",
        "operationId": "ColleagueGet",
        "parameters": [
          {
            "name": "Filters",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "Sorts",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 2147483647.0,
              "minimum": 1.0,
              "nullable": true
            },
            "x-position": 3
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 2147483647.0,
              "minimum": 1.0,
              "nullable": true
            },
            "x-position": 4
          },
          {
            "name": "withIncludes",
            "in": "query",
            "description": "Specifies whether related entities should be included in the query.",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "x-position": 5
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ColleagueDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "Admin",
              " User"
            ]
          }
        ]
      },
      "post": {
        "tags": [
          "Colleague"
        ],
        "summary": "Creates a new colleague.",
        "operationId": "ColleagueCreate",
        "requestBody": {
          "x-name": "payload",
          "description": "The colleague data.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ColleagueCreateDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateActionResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "Admin"
            ]
          }
        ]
      }
    },
    "/api/v1/colleagues/{id}": {
      "get": {
        "tags": [
          "Colleague"
        ],
        "summary": "Retrieves a colleague by its ID.",
        "operationId": "ColleagueGetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the colleague to retrieve.",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ColleagueDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "Admin",
              " User"
            ]
          }
        ]
      },
      "patch": {
        "tags": [
          "Colleague"
        ],
        "summary": "Updates a colleague by ID.",
        "operationId": "ColleagueUpdate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the colleague to update.",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "patchDocument",
          "description": "The JSON patch document containing updates.",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Operation"
                }
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "Admin"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Colleague"
        ],
        "summary": "Deletes a colleague by ID.",
        "operationId": "Colleague_ColleagueDelete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the colleague to delete.",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "Admin"
            ]
          }
        ]
      }
    },
    "/api/v1/employees": {
      "get": {
        "tags": [
          "Employee"
        ],
        "summary": "Retrieves a list of employees.",
        "operationId": "EmployeeGet",
        "parameters": [
          {
            "name": "Filters",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "Sorts",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 2147483647.0,
              "minimum": 1.0,
              "nullable": true
            },
            "x-position": 3
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 2147483647.0,
              "minimum": 1.0,
              "nullable": true
            },
            "x-position": 4
          },
          {
            "name": "withIncludes",
            "in": "query",
            "description": "Specifies whether related entities should be included in the query.",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "x-position": 5
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmployeeDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "Admin",
              " User"
            ]
          }
        ]
      },
      "post": {
        "tags": [
          "Employee"
        ],
        "summary": "Creates new employee.",
        "operationId": "EmployeeCreate",
        "requestBody": {
          "x-name": "payload",
          "description": "The employee content.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeCreateDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "The result of creation. CreateActionResultDto",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateActionResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "Admin"
            ]
          }
        ]
      }
    },
    "/api/v1/employees/{id}": {
      "get": {
        "tags": [
          "Employee"
        ],
        "summary": "Retrieves a employee by its ID.",
        "operationId": "EmployeeGetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the employee to retrieve.",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "Admin",
              " User"
            ]
          }
        ]
      },
      "patch": {
        "tags": [
          "Employee"
        ],
        "summary": "Updates a employee by ID.",
        "operationId": "EmployeeUpdate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the employee to update.",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "patchDocument",
          "description": "The JSON patch document containing updates.",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Operation"
                }
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "Admin"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Employee"
        ],
        "summary": "Deletes a employee by ID.",
        "operationId": "EmployeeDelete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the employee to delete.",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "Admin",
              " User"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ColleagueDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DtoBase"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "employeeId",
              "colleagueEmployeeId"
            ],
            "properties": {
              "employeeId": {
                "type": "string",
                "format": "guid",
                "minLength": 1
              },
              "colleagueEmployeeId": {
                "type": "string",
                "format": "guid",
                "minLength": 1
              }
            }
          }
        ]
      },
      "DtoBase": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "guid",
            "minLength": 1
          }
        }
      },
      "ColleagueDetailDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DtoBase"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "colleagueEmployee"
            ],
            "properties": {
              "colleagueEmployee": {
                "$ref": "#/components/schemas/EmployeeDetailDto"
              }
            }
          }
        ]
      },
      "EmployeeDetailDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DtoBase"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "name",
              "surname",
              "email",
              "position"
            ],
            "properties": {
              "name": {
                "type": "string",
                "minLength": 1
              },
              "surname": {
                "type": "string",
                "minLength": 1
              },
              "patronymic": {
                "type": "string",
                "nullable": true
              },
              "email": {
                "type": "string",
                "minLength": 1
              },
              "position": {
                "$ref": "#/components/schemas/PositionDto"
              },
              "colleagues": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ColleagueDetailDto"
                }
              }
            }
          }
        ]
      },
      "PositionDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DtoBase"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "name"
            ],
            "properties": {
              "name": {
                "type": "string",
                "minLength": 1
              },
              "description": {
                "type": "string",
                "nullable": true
              }
            }
          }
        ]
      },
      "ErrorDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "status"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "stackTrace": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "extensions": {
            "type": "object",
            "additionalProperties": {}
          }
        }
      },
      "CreateActionResultDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "guid",
            "minLength": 1
          }
        }
      },
      "ColleagueCreateDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "employeeId",
          "colleagueEmployeeId"
        ],
        "properties": {
          "employeeId": {
            "type": "string",
            "format": "guid",
            "minLength": 1
          },
          "colleagueEmployeeId": {
            "type": "string",
            "format": "guid",
            "minLength": 1
          }
        }
      },
      "Operation": {
        "allOf": [
          {
            "$ref": "#/components/schemas/OperationBase"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "value": {
                "nullable": true
              }
            }
          }
        ]
      },
      "OperationBase": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "path": {
            "type": "string",
            "nullable": true
          },
          "op": {
            "type": "string",
            "nullable": true
          },
          "from": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "EmployeeDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DtoBase"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "name",
              "surname",
              "position"
            ],
            "properties": {
              "name": {
                "type": "string",
                "minLength": 1
              },
              "surname": {
                "type": "string",
                "minLength": 1
              },
              "patronymic": {
                "type": "string",
                "nullable": true
              },
              "position": {
                "$ref": "#/components/schemas/PositionDto"
              }
            }
          }
        ]
      },
      "EmployeeCreateDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "surname",
          "email",
          "password",
          "positionId"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "surname": {
            "type": "string",
            "minLength": 1
          },
          "patronymic": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 1
          },
          "positionId": {
            "type": "string",
            "format": "guid",
            "minLength": 1
          }
        }
      },
...
    },
    "securitySchemes": {
      "JWT": {
        "type": "apiKey",
        "description": "Type into the textbox: Bearer {your JWT token}.",
        "name": "Authorization",
        "in": "header",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}